tf_serving/requirements.txt
keras-image-helper
tensorflow-protobuf
uvicorn
grpcio
fastapi